Ignore match_like_matches_macro lint for 'scalar()' generated function.
The functions may be generated with a match block similar to:

match self {
   Value::Parameter(_) => true,
    _ => false,
}

This pattern is what the matches! macro was intended to simplify.
Detecting this edge case would require additional state in the code
generation logic that's undesirable. The simple solution is to have this
lint ignore these functions.