Ignore the 'should_implement_trait' lint for 'from_str' function.
This lint is triggered because the FromStr interface has a function with
the same name. Rust believes that developers will see the 'from_str'
function and assume that the type must implement the FromStr trait when
in fact it hasn't. The code generated here isn't really a public API so
these concerns are mostly a false positive. Additionally the 'from_str'
function from the FromStr trait has a different signature so
implementing this interface would be more trouble than it's worth.