clippy fixes (#4)
* Pass Vecs as slices instead of concrete types.

* Replace match with one arm with if let.

* In loops that ingnore values, iterate over keys only.

* Remove unnecessary borrows.

These are immediately dereferenced by the compiler according to clippy.

* Simplify match with single pattern to if let.

* Use a type suffix on literals instead of casting with 'as'.

* Convert enum variant with an all uppercase name to UpperCamelCase.

This gets caught by the upper-case-acronyms clippy lint.

* Replace 'if let' with 'is_err()' utility function.

* Ignore the match_single_binding clippy lint for some generated functions.

The functions may be generated as:

match self {
    _ => false,
}

that could be simplified to: false. Unfortunately it looks like
detecting this edge case would require additional state in the code
generation logic and that additional complexity is undesirable. The
simple solution is to have this lint ignore these functions.

* Ignore match_like_matches_macro lint for 'scalar()' generated function.

The functions may be generated with a match block similar to:

match self {
   Value::Parameter(_) => true,
    _ => false,
}

This pattern is what the matches! macro was intended to simplify.
Detecting this edge case would require additional state in the code
generation logic that's undesirable. The simple solution is to have this
lint ignore these functions.

* Ignore the 'should_implement_trait' lint for 'from_str' function.

This lint is triggered because the FromStr interface has a function with
the same name. Rust believes that developers will see the 'from_str'
function and assume that the type must implement the FromStr trait when
in fact it hasn't. The code generated here isn't really a public API so
these concerns are mostly a false positive. Additionally the 'from_str'
function from the FromStr trait has a different signature so
implementing this interface would be more trouble than it's worth.

* Ignore useless conversions from '.into()' and 'u32::try_from()'.

This is more complex than removing unnecessary '.into()' and
'u32::try_from()' calls because in some cases the code generated
requires them. Ignoring the cases where they're unnecessary keeps us
from further complicating the code generation.

* Ignore identity_op lint for 'to_slice' functions.

Generating this function for tuple structs with a single u8 will produce
an ineffective operation ((self.0 >> 0) & 0xff). Accounting for this
edge case would add complexity to the code generation logic so ignoring
the lint for these functions seems preferable.