Make build output reproducible
The build code iterates over HashMaps to generate output. This is
correct but may not always generate the same output depending on
hashing. This can produce slightly different generated code.
Ensure the output is stable by sorting before outputing. We don't
actually care what the order is, just that it happens the same way
every time.