fix invalid sentinel values + ADM1272 OPERATION definition (#20)
When invalid sentinel values are discovered in the results of a PMBus
command, the behavior has historically been to indicate an error for
anyone trying to interpret it.  This behavior has proved to be too
persnickety: in practice, it's hard for the caller to really do
anything with the error, especially because there is no indication of
which field was found to have an illegal sentinel or what that
sentinel was found to be.  So instead of indicating an InvalidSentinel
error, this change uses our existing Unknown variants for invalid
sentinels -- which allows for processing to complete on other values,
and also makes clear that the sentinal value is unknown.  At the same
time, this work adds the OPERATION definition for the ADM1272,
a(nother) part that falls into exactly the trap that this work
alleviates.